papercoder/requirements.txt
# ── Agent 编排 ──────────────────────────────────────────────────
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-anthropic>=0.3.0
langchain-openai>=0.2.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0
langsmith>=0.1.0

# ── 学术检索 ─────────────────────────────────────────────────────
arxiv>=2.1.0
requests>=2.31.0

# ── 网络搜索 ─────────────────────────────────────────────────────
tavily-python>=0.3.0

# ── 本地 RAG ─────────────────────────────────────────────────────
llama-index>=0.11.0
llama-index-vector-stores-chroma>=0.2.0
chromadb>=0.5.0

# ── PDF 解析 ─────────────────────────────────────────────────────
pymupdf>=1.24.0
# marker-pdf>=0.2.0        # 可选：高质量结构化解析（安装较慢）
# nougat-ocr>=0.1.0        # 可选：数学公式专项

# ── 长期记忆 ─────────────────────────────────────────────────────
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0
numpy>=1.24.0

# ── MCP ─────────────────────────────────────────────────────────
mcp>=1.0.0

# ── 工具类 ──────────────────────────────────────────────────────
pydantic>=2.0.0
tenacity>=8.0.0
python-dotenv>=1.0.0
